refactor(server): JS can import and run with one line
When import as a module, it runs automatically.
